UserProfiles controller is finished. Working on posts controller

+++ CwkSocial.Application/UserProfiles/CommandHandlers/DeleteUserProfileHandler.cs
@@ -1,4 +1,7 @@
-﻿using CwkSocial.Application.UserProfiles.Commands;
+﻿using Cwk.Domain.Aggregates.UserProfileAggregate;
+using CwkSocial.Application.Enums;
+using CwkSocial.Application.Models;
+using CwkSocial.Application.UserProfiles.Commands;
 using CwkSocial.Dal;
 using MediatR;
 using Microsoft.EntityFrameworkCore;
@@ -11,25 +14,39 @@ using System.Threading.Tasks;
 
 namespace CwkSocial.Application.UserProfiles.CommandHandlers
 {
-    internal class DeleteUserProfileHandler : IRequestHandler<DeleteUserProfile>
+    internal class DeleteUserProfileHandler : IRequestHandler<DeleteUserProfile, OperationResult<UserProfile>>
     {
         private readonly DataContext _context;
         public DeleteUserProfileHandler(DataContext context)
         {
             _context = context;
         }
-        public async Task Handle(DeleteUserProfile request, CancellationToken cancellationToken)
+        public async Task<OperationResult<UserProfile>> Handle(DeleteUserProfile request, CancellationToken cancellationToken)
         {
+            var result = new OperationResult<UserProfile>();
+
             var userProfile = 
                 await _context.UserProfiles.FirstOrDefaultAsync(up => up.UserProfileId == request.UserProfileId);
 
-            //if (userProfile == null)
-            //    return new Unit();
+            if (userProfile == null)
+            {
+                result.IsError = true;
+                var error = new Error
+                {
+                    Code = ErrorCode.NotFound,
+                    Message = $"No UserProfile found with ID {request.UserProfileId}"
+                };
+                result.Errors.Add(error);
+
+                return result;
+            }                
 
             _context.UserProfiles.Remove(userProfile);
             await _context.SaveChangesAsync();
 
-           
+            result.Payload = userProfile;
+
+           return result;
         }
     }
 }
